pyproject.toml
[project]
name = "sanitarskyi-tg-coding-assistant"
version = "0.1.0"
description = "Telegram bot for Cursor Cloud Agent API"
requires-python = ">=3.10,<3.14"
dependencies = [
    "aiogram==3.13.1",
    "httpx==0.27.0",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.23.7",
    "pytest-mock==3.14.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
version = { source = "file", path = "pyproject.toml" }

[tool.pdm.scripts]
start = "python main.py"
test = "pytest"
lint = "ruff check ."
format = "sh -c 'black . && isort .'"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["bot", "cursor"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

